requirements.txt
openai>=1.0.0
PyPDF2>=3.0.0
flask>=3.0.0
requests>=2.31.0
markdown>=3.5.0
bleach>=6.0.0
python-dotenv>=1.0.0
